Make some dubious type conversions explicit.
This is to address MSVC warnings.

This commit changes a few configuration fields to use double instead
of float. There doesn't seem to be any reason these use float except
for the legacy Windows code using float for saved configuration.
Changing their type to double improves consistency.

+++ src/mouse.cpp
@@ -909,7 +909,7 @@ bool GraphicsWindow::MouseEvent(Platform::MouseEvent event) {
             break;
 
         case MouseEvent::Type::SCROLL_VERT:
-            this->MouseScroll(event.x, event.y, event.scrollDelta);
+            this->MouseScroll(event.x, event.y, (int)event.scrollDelta);
             break;
 
         case MouseEvent::Type::LEAVE:
